Manipulator's functions have been implemented

+++ src/util/collections/ArduinoList.h
@@ -1,20 +1,22 @@
 #pragma once
 
+#include <stdint.h>
+
 template <typename T>
 class ArduinoList {
 private:
-    size_t arraySize = 0;
+    uint16_t arraySize = 0;
     T* data = nullptr;
 public:
-    T& operator [](size_t id) {
+    T& operator [](uint16_t id) {
         return data[id];
     }
 
-    void removeAt(size_t id) {
+    void removeAt(uint16_t id) {
         arraySize--;
         if (arraySize > 0) {
             T* newData = new T[arraySize];
-            for (size_t elementId = 0, newArrayId = 0; elementId < arraySize + 1; elementId++) {
+            for (uint16_t elementId = 0, newArrayId = 0; elementId < arraySize + 1; elementId++) {
                 if (elementId != id) {
                     newData[newArrayId++] = data[elementId];
                 }
@@ -28,7 +30,7 @@ public:
     }
 
     void remove(T element) {
-        for (size_t elementId = 0; elementId < arraySize; elementId++)
+        for (uint16_t elementId = 0; elementId < arraySize; elementId++)
         {
             if (data[elementId] == elementId) {
                 removeAt(elementId);
@@ -40,7 +42,7 @@ public:
         arraySize++;
         T* newData = new T[arraySize];
         if (data != nullptr) {
-            for (size_t elementId = 0; elementId < arraySize - 1; elementId++) {
+            for (uint16_t elementId = 0; elementId < arraySize - 1; elementId++) {
                 newData[elementId] = data[elementId];
             }
             delete[] data;
@@ -55,7 +57,7 @@ public:
         return *this;
     } 
 
-    size_t size() {
+    uint16_t size() {
         return this->arraySize;
     }
 
@@ -64,7 +66,7 @@ public:
     ArduinoList(const ArduinoList<T>& list) {
         this->arraySize = list.arraySize;
         this->data = new T[this->arraySize];
-        for (size_t elementId = 0; elementId < list.arraySize; elementId++)
+        for (uint16_t elementId = 0; elementId < list.arraySize; elementId++)
         {
             data[elementId] = list.data[elementId];
         }
